# lib/api-zod/src/generated/types/renderInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Autonomous YouTube Factory API
 * OpenAPI spec version: 1.0.0
 */
import type { RenderInputAspectRatio } from './renderInputAspectRatio';
import type { RenderInputCropMode } from './renderInputCropMode';
import type { RenderInputResolution } from './renderInputResolution';

export interface RenderInput {
  timelineId: string;
  /** @nullable */
  voiceId?: string | null;
  resolution?: RenderInputResolution;
  /**
     * @minimum 24
     * @maximum 60
     */
  fps?: number;
  aspectRatio?: RenderInputAspectRatio;
  cropMode?: RenderInputCropMode;
  hardwareAcceleration?: boolean;
  addBackgroundMusic?: boolean;
  /**
     * @minimum 0
     * @maximum 1
     */
  musicVolume?: number;
  generatePreview?: boolean;
}