# lib/api-zod/src/generated/types/renderList.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Autonomous YouTube Factory API
 * OpenAPI spec version: 1.0.0
 */
import type { RenderResult } from './renderResult';

export interface RenderList {
  items: RenderResult[];
  total: number;
}