# lib/api-zod/src/generated/types/renderMetadata.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Autonomous YouTube Factory API
 * OpenAPI spec version: 1.0.0
 */

export interface RenderMetadata {
  sceneCount?: number;
  clipCount?: number;
  placeholderClipCount?: number;
  hasNarration?: boolean;
  hasBackgroundMusic?: boolean;
  sourceTimelineId?: string;
  /** @nullable */
  sourceVoiceId?: string | null;
}