# lib/api-zod/src/generated/types/renderOutput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Autonomous YouTube Factory API
 * OpenAPI spec version: 1.0.0
 */

export interface RenderOutput {
  /** @nullable */
  localPath?: string | null;
  fileSizeBytes?: number;
  durationSeconds?: number;
  width?: number;
  height?: number;
  fps?: number;
  codec?: string;
  audioCodec?: string;
  format?: string;
}