# lib/api-zod/src/generated/types/renderResult.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Autonomous YouTube Factory API
 * OpenAPI spec version: 1.0.0
 */
import type { RenderMetadata } from './renderMetadata';
import type { RenderOutput } from './renderOutput';
import type { RenderResultRenderPlan } from './renderResultRenderPlan';
import type { RenderResultStatus } from './renderResultStatus';
import type { RenderStats } from './renderStats';

export interface RenderResult {
  id: string;
  timelineId: string;
  /** @nullable */
  voiceId?: string | null;
  status: RenderResultStatus;
  progress?: number;
  resolution?: string;
  width?: number;
  height?: number;
  fps?: number;
  aspectRatio?: string;
  cropMode?: string;
  hardwareAcceleration?: boolean;
  renderPlan?: RenderResultRenderPlan;
  renderOutput?: RenderOutput;
  renderStats?: RenderStats;
  renderMetadata?: RenderMetadata;
  previewOutput?: RenderOutput;
  logs: string[];
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  completedAt?: string | null;
}