# lib/api-zod/src/generated/types/renderStats.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Autonomous YouTube Factory API
 * OpenAPI spec version: 1.0.0
 */

export interface RenderStats {
  renderTimeSeconds?: number;
  framesEncoded?: number;
  encodeFps?: number;
  realtimeFactor?: number;
  retries?: number;
}